refactor(images): Removed IdImageProvider
Can now specify directly on each region.
