+++ keel-ec2-plugin/src/main/kotlin/com/netflix/spinnaker/keel/ec2/jackson/ImageProviderDeserializer.kt
@@ -19,7 +19,6 @@ package com.netflix.spinnaker.keel.ec2.jackson
 
 import com.netflix.spinnaker.keel.api.InvalidPayload
 import com.netflix.spinnaker.keel.api.ec2.ArtifactImageProvider
-import com.netflix.spinnaker.keel.api.ec2.IdImageProvider
 import com.netflix.spinnaker.keel.api.ec2.ImageProvider
 import com.netflix.spinnaker.keel.api.ec2.JenkinsImageProvider
 import com.netflix.spinnaker.keel.serialization.PropertyNamePolymorphicDeserializer
@@ -28,7 +27,6 @@ internal class ImageProviderDeserializer :
   PropertyNamePolymorphicDeserializer<ImageProvider>(ImageProvider::class.java) {
   override fun identifySubType(fieldNames: Collection<String>): Class<out ImageProvider> =
     when {
-      "imageId" in fieldNames -> IdImageProvider::class.java
       "deliveryArtifact" in fieldNames -> ArtifactImageProvider::class.java
       "buildHost" in fieldNames -> JenkinsImageProvider::class.java
       else -> throw InvalidPayload("ImageProvider")
